+++ mkwutil/lib/ppc_dis.py
