+++ mkwutil/lib/slices_test.py
