+++ mkwutil/lib/symbols.py
