+++ mkwutil/tools/dump_elf_segments.py
