+++ mkwutil/tools/dump_symbols.py
