+++ mkwutil/tools/format_symbols.py
